### 2022-06-15 3e2b7a4
A few minor paper fixes
