+++ paper/paper.md
@@ -57,7 +57,7 @@ through a SWIG interface to LALSuite [@WETTE2020100634]. The Bayesian inference 
 the [`bilby`](https://lscsoft.docs.ligo.org/bilby/) package [@2019ApJS24127A], which provides a
 convenient interface to a wide variety of packages for using the Markov Chain Monte Carlo (MCMC;
 see, e.g.\ @2017ARA&A55213S for a review of MCMC with particular reference to astronomy) or nested
-sampling algorithms (see, e.g., @2021arXiv210109675B or @nestedsampling for a review). By default
+sampling algorithms (see, e.g., @2021arXiv210109675B or @nestedsampling for a review). By default,
 CWInPy uses the [`dynesty`](https://dynesty.readthedocs.io/) package [@2020MNRAS.493.3132S] for
 inference using nested sampling, producing both posterior probability distributions for the
 parameters of interest and the Bayesian evidence for the signal model given the data.
@@ -66,8 +66,8 @@ parameters of interest and the Bayesian evidence for the signal model given the 
 
 CWInPy is designed to supersede the current analysis pipeline, known as `lalapps_knope`
 [@2017arXiv170508978P], largely based on executables written in `C`, which has been used for several
-searches in LIGO and Virgo data [@2017ApJ83912A; @2019ApJ87910A; @2020ApJ902L21A]. The reasons
-behind CWInPy's development, and its enhancements over existed software, include:
+searches in LIGO and Virgo data [@2017ApJ83912A; @2019ApJ87910A; @2020ApJ902L21A; @2021arXiv211113106T].
+The reasons behind CWInPy's development, and its enhancements over existed software, include:
 
 * the Python API allows easy access to the full range of functionality from data preprocessing to
   source parameter estimation, with greater ability to control various aspects of the analysis;
@@ -93,20 +93,20 @@ behind CWInPy's development, and its enhancements over existed software, include
   a population of pulsars [@2018PhRvD98f3001P];
 * it provides tools for determining the significance of a recovered signal using the sky-shifting
   method [@2020PhRvD.102l3027I];
-* the pulsar phase evolution can be calculated using the standard pulsar timing package TEMPO2
+* the pulsar phase evolution can be calculated using the standard pulsar timing package Tempo2
   [@tempo2] via the [`libstempo`](https://github.com/vallis/libstempo) Python package
   [@2020ascl.soft02017V] in addition to being able to use the independent routines within LALSuite
   [@lalsuite].
 
 # Validation
 
-To validate CWInPy, extensive testing against previous codes has been performed and is provided with
+To validate CWInPy extensive testing against previous codes has been performed and is provided with
 the [documentation](https://cwinpy.readthedocs.io/en/latest/). CWInPy has been used to successfully
 extract simulated signals directly added into the LIGO data [@2017PhRvD95f2002B] as well as multiple
 signals simulated via software.
 
 The inference of source parameters has been validated through the use of simulation-based
-calibration [@2018arXiv180406788T], with the posteriors probability distributions provided
+calibration [@2018arXiv180406788T], with the posteriors probability distributions shown to provide
 well-calibrated [@wellcalibrated] credible intervals.
 
 As evidence for this consistency, \autoref{fig:posteriors} shows the posterior probability
@@ -148,9 +148,9 @@ Both the `cwinpy_knope_pipeline` and the `cwinpy_heterodyne_pipeline` executable
 command line arguments that can be used for quickly setting up analyses using open data from GWOSC.
 These rely on the user having access to a computer, or cluster of computers, with HTCondor installed
[…]
