### 2022-06-17 #4476 @arfon
:wave: @mattpitkin – thanks for your submission to JOSS. We're a little short on editorial capacity to handle this submission right now so I'm going to waitlist this until we have capacity (this will likely be less than a month).

Thanks in advance for your patience!
