### 2022-06-17 ac0c2b0
Start adding documentation for an example using the transient model
