+++ docs/pe/data/example3_config.ini
@@ -1,0 +1,32 @@
+# configuration file for Example 3
+
+# The path to the TEMPO(2)-style pulsar parameter file
+par-file=TRANSIENT.par
+
+# The paths to the TEMPO(2)-style injection file
+inj-file=TRANSIENT.par
+
+# The start and end times of the simulated data
+fake-start={'H1': 1000000000, 'L1': 1000000000}
+fake-end={'H1': 1000345600, 'L1': 1000345600}
+
+# Use design curve ASD for simulated noise
+fake-asd=[H1, L1]
+
+# The output directory for the results (this will be created if it does not exist)
+outdir=example3
+
+# A prefix for the results file name
+label=example3
+
+# The Bayesian stochastic sampling algorithm (defaults to dynesty if not given)
+sampler=dynesty
+
+# Keyword arguments for the sampling algorithm
+sampler-kwargs={'Nlive': 1000, 'plot': True}
+
+# Show "true" signal values on the output plot (as we know this data contains a simulated signal!)
+show-truths = True
+
+# The path to the prior file
+prior=example3_prior.txt
