+++ docs/pe/data/example3_prior.txt
@@ -4,3 +4,5 @@ h0 = Uniform(name='h0', minimum=0, maximum=1e-22, latex_label='$h_0$')
 iota = Sine(name='iota', minimum=0., maximum=np.pi, latex_label='$\iota$', unit='rad')
 phi0 = Uniform(name='phi0', minimum=0, maximum=np.pi, latex_label='$\phi_0$', unit='rad')
 psi = Uniform(name='psi', minimum=0, maximum=np.pi / 2, latex_label='$\psi$', unit='rad')
+transientstarttime = Gaussian(name='transientstarttime', mu=1000043200, sigma=43200, latex_label='$t_0$', unit='s')
+transienttau = Uniform(name='transienttau', minimum=8640, maximum=259200, latex_label='$\\tau$', unit='s')
