+++ docs/pe/data/example3_config.ini
@@ -1,14 +1,12 @@
 # configuration file for Example 3
 
-# The path to the TEMPO(2)-style pulsar parameter file
-par-file=TRANSIENT.par
-
 # The paths to the TEMPO(2)-style injection file
-inj-file=TRANSIENT.par
+par-file=TRANSIENT.par
+inj-par=TRANSIENT.par
 
 # The start and end times of the simulated data
-fake-start={'H1': 1000000000, 'L1': 1000000000}
-fake-end={'H1': 1000345600, 'L1': 1000345600}
+fake-start=1000000000
+fake-end=1000345600
 
 # Use design curve ASD for simulated noise
 fake-asd=[H1, L1]
