+++ docs/knope/examples/knope_example_O2_vela_transient.ini
@@ -1,0 +1,35 @@
+[run]
+basedir = /home/matthew.pitkin/projects/cwinpyO2
+
+[knope_job]
+accounting_group = ligo.prod.o2.cw.targeted.bayesian
+accounting_group_user = matthew.pitkin
+
+[heterodyne_job]
+getenv = True
+
+[ephemerides]
+pulsarfiles = /home/matthew.pitkin/projects/O2pulsars
+
+[heterodyne]
+detectors = ["H1", "L1"]
+freqfactors = 2
+starttimes = 1164556817
+endtimes = 1187733618
+frametypes = {"H1": "H1_GWOSC_O2_4KHZ_R1", "L1": "L1_GWOSC_O2_4KHZ_R1"}
+channels = {"H1":"H1:GWOSC-4KHZ_R1_STRAIN", "L1": "L1:GWOSC-4KHZ_R1_STRAIN"}
+host = datafind.gw-openscience.org
+includeflags = {"H1": "H1_CBC_CAT1", "L1": "L1_CBC_CAT1"}
+outputdir = {"H1": "/home/matthew.pitkin/projects/cwinpyO2/heterodyne/H1", "L1": "/home/matthew.pitkin/projects/cwinpyO2/heterodyne/L1"}
+usetempo2 = False
+
+[merge]
+remove = True
+
+[pe_dag]
+submitdag = True
+
+[pe]
+priors = /home/matthew.pitkin/projects/O2priors/vela_transient_prior.txt
+incoherent = True
+
