### 2022-06-22 de59176
Fix a couple of typos
