+++ docs/pe/data/example3_config.ini
@@ -15,7 +15,7 @@ fake-asd=[H1, L1]
 outdir=example3
 
 # A prefix for the results file name
-label=example3_update
+label=example3
 
 # The Bayesian stochastic sampling algorithm (defaults to dynesty if not given)
 sampler=dynesty
@@ -27,4 +27,4 @@ sampler-kwargs={'Nlive': 1000, 'plot': True}
 show-truths = True
 
 # The path to the prior file
-prior=example3_prior_update.txt
+prior=example3_prior.txt
