+++ docs/knope/examples/knope_example_O2_vela_transient.ini
@@ -1,5 +1,5 @@
 [run]
-basedir = /home/matthew.pitkin/projects/cwinpyO2
+basedir = /home/matthew.pitkin/projects/vela
 
 [knope_job]
 accounting_group = ligo.prod.o2.cw.targeted.bayesian
@@ -9,7 +9,7 @@ accounting_group_user = matthew.pitkin
 getenv = True
 
 [ephemerides]
-pulsarfiles = /home/matthew.pitkin/projects/O2pulsars
+pulsarfiles = /home/matthew.pitkin/projects/vela/vela.par
 
 [heterodyne]
 detectors = ["H1", "L1"]
@@ -20,7 +20,7 @@ frametypes = {"H1": "H1_GWOSC_O2_4KHZ_R1", "L1": "L1_GWOSC_O2_4KHZ_R1"}
 channels = {"H1":"H1:GWOSC-4KHZ_R1_STRAIN", "L1": "L1:GWOSC-4KHZ_R1_STRAIN"}
 host = datafind.gw-openscience.org
 includeflags = {"H1": "H1_CBC_CAT1", "L1": "L1_CBC_CAT1"}
-outputdir = {"H1": "/home/matthew.pitkin/projects/cwinpyO2/heterodyne/H1", "L1": "/home/matthew.pitkin/projects/cwinpyO2/heterodyne/L1"}
+outputdir = {"H1": "/home/matthew.pitkin/projects/vela/heterodyne/H1", "L1": "/home/matthew.pitkin/projects/vela/heterodyne/L1"}
 usetempo2 = False
 
 [merge]
@@ -30,6 +30,6 @@ remove = True
 submitdag = True
 
 [pe]
-priors = /home/matthew.pitkin/projects/O2priors/vela_transient_prior.txt
+priors = /home/matthew.pitkin/projects/vela/vela_transient_prior.txt
 incoherent = True
 
