### 2022-06-24 7f198f0
Remove spurious blank line from config file
