+++ docs/knope/examples/knope_example_O2_vela_transient.ini
@@ -32,4 +32,3 @@ submitdag = True
 [pe]
 priors = /home/matthew.pitkin/projects/vela/vela_transient_prior.txt
 incoherent = True
-
