### 2022-06-24 139aff6
Update build.yml: remove setup python step as we're using a container
