### 2022-06-24 c667e5d
Update comparison plots
