+++ docs/comparisons/scripts/multi_detector_noise_only.py
@@ -7,21 +7,19 @@ data for a multiple detectors (H1, L1 and V1).
 
 import os
 import subprocess as sp
-from collections import OrderedDict
 
-import corner
 import h5py
 import matplotlib
-import matplotlib.font_manager as font_manager
 import numpy as np
 from astropy.utils.data import download_file
 from bilby.core.prior import Uniform
 from comparitors import comparisons
 from cwinpy import HeterodynedData
 from cwinpy.pe import pe
+from cwinpy.plot import Plot
 from lalinference import LALInferenceHDF5PosteriorSamplesDatasetName
 from lalinference.io import read_samples
-from matplotlib.lines import Line2D
+from matplotlib import pyplot as plt
 
 matplotlib.use("Agg")
 
@@ -36,10 +34,6 @@ DECJ     34:56:54.321
 F0       567.89
 F1       -1.2e-12
 PEPOCH   56789
-H0       5.12e-25
-COSIOTA  0.3
-PSI      1.1
-PHI0     2.4
 """
 
 label = "multi_detector_noise_only"
@@ -86,7 +80,7 @@ with open(priorfile, "w") as fp:
     fp.write(priorcontent.format(*(h0range + phi0range + psirange + cosiotarange)))
 
 # set prior for bilby
-priors = OrderedDict()
+priors = {}
 priors["h0"] = Uniform(h0range[0], h0range[1], "h0", latex_label=r"$h_0$")
 priors["phi0"] = Uniform(
     phi0range[0], phi0range[1], "phi0", latex_label=r"$\phi_0$", unit="rad"
@@ -201,7 +195,7 @@ runner = pe(
 result = runner.result
 
 # evaluate the likelihood on a grid
-gridpoints = 35
+gridpoints = 30
 grid_size = dict()
 for p in priors.keys():
     grid_size[p] = np.linspace(
@@ -224,40 +218,31 @@ grid = grunner.grid
 # output comparisons
 comparisons(label, outdir, grid, priors, cred=0.9)
 
-# plot results
-fig = result.plot_corner(save=False, parameters=list(priors.keys()), color="b")
-fig = corner.corner(
-    postsamples,
-    fig=fig,
-    color="r",
+# create results plot
+allresults = {
+    "lalapps_pulsar_parameter_estimation_nested": outpost,
+    "cwinpy_pe": result,
+    "cwinpy_pe (grid)": grid,
+}
+
+colors = {
+    key: plt.rcParams["axes.prop_cycle"].by_key()["color"][i]
+    for i, key in enumerate(allresults.keys())
+}
+
+plot = Plot(
+    results=allresults,
+    parameters=list(priors.keys()),
+    plottype="corner",
+)
+
+plot.plot(
     bins=50,
     smooth=0.9,
     quantiles=[0.16, 0.84],
     levels=(1 - np.exp(-0.5), 1 - np.exp(-2), 1 - np.exp(-9 / 2.0)),
     fill_contours=True,
-    hist_kwargs={"density": True},
+    colors=colors,
 )
-axes = fig.get_axes()
-axidx = 0
-for p in priors.keys():
-    axes[axidx].plot(
-        grid.sample_points[p],
-        np.exp(grid.marginalize_ln_posterior(not_parameters=p) - grid.log_evidence),
-        "k--",
-    )
-    axidx += 5
 
-# custom legend
-legend_elements = [
-    Line2D([], [], color="r", label="lalapps_pulsar_parameter_estimation_nested"),
-    Line2D([], [], color="b", label="cwinpy_pe"),
-    Line2D([], [], color="k", ls="--", label="cwinpy_pe (grid)"),
-]
-font = font_manager.FontProperties(family="monospace")
[…]
