+++ setup.cfg
@@ -37,6 +37,7 @@ install_requires =
     corner >= 2.2.1
     gwosc >= 0.5.4
     gwpy >= 2.1.1
+    healpy
     htcondor >= 9.0.6
     lalsuite >= 7.5.1.dev20220614
     lintegrate >= 0.1.7
@@ -64,7 +65,6 @@ test =
     coverage
     coverage-badge
     cweqgen >= 0.4.3
-    healpy
     libstempo >= 2.4.2
     pytest >= 4.6
     pytest_socket

[…]
