### 2022-06-24 4bd7b0e
Update container location for CI
