+++ docs-environment.yml
@@ -1,154 +1,39 @@
 name: cwinpydocs
 channels:
   - conda-forge
-  - defaults
 dependencies:
-  - _libgcc_mutex=0.1=conda_forge
-  - _openmp_mutex=4.5=1_gnu
-  - ca-certificates=2020.12.5=ha878542_0
-  - certifi=2020.12.5=py37h89c1867_1
-  - gsl=2.6=he838d99_2
-  - ld_impl_linux-64=2.35.1=hea4e1c9_2
-  - libblas=3.9.0=8_openblas
-  - libcblas=3.9.0=8_openblas
-  - libffi=3.3=h58526e2_2
-  - libgcc-ng=9.3.0=h2828fa1_18
-  - libgfortran-ng=9.3.0=hff62375_18
-  - libgfortran5=9.3.0=hff62375_18
-  - libgomp=9.3.0=h2828fa1_18
-  - liblapack=3.9.0=8_openblas
-  - libopenblas=0.3.12=pthreads_h4812303_1
-  - libstdcxx-ng=9.3.0=h6de172a_18
-  - lintegrate=0.1.7=py37h219ec46_0
-  - ncurses=6.2=h58526e2_4
-  - numpy=1.19.5=py37haa41c4c_1
-  - openssl=1.1.1j=h7f98852_0
-  - pip=21.0.1=pyhd8ed1ab_0
-  - python=3.7.10=hffdb5ce_100_cpython
-  - python_abi=3.7=1_cp37m
-  - readline=8.0=he28a2e2_2
-  - setuptools=49.6.0=py37h89c1867_3
-  - sqlite=3.34.0=h74cdb3f_0
-  - tk=8.6.10=h21135ba_1
-  - wheel=0.36.2=pyhd3deb0d_0
-  - xz=5.2.5=h516909a_1
-  - zlib=1.2.11=h516909a_1010
+  - python=3.8
   - pip:
-    - alabaster==0.7.12
-    - appdirs==1.4.4
-    - astropy==4.2
-    - async-generator==1.10
-    - attrs==20.3.0
-    - autodoc==0.5.0
-    - babel==2.9.0
-    - beautifulsoup4==4.9.3
-    - bilby==1.1.4
-    - bilby-pipe==1.0.2
-    - bleach==3.3.0
-    - cached-property==1.5.2
-    - cffi==1.14.5
-    - chardet==4.0.0
-    - click==7.1.2
-    - commonmark==0.9.1
-    - configargparse==1.3
-    - corner==2.2.1
-    - cryptography==3.3.1
-    - cycler==0.10.0
-    - decorator==4.4.2
-    - deepdish==0.3.6
-    - defusedxml==0.6.0
-    - dill==0.3.3
-    - docutils==0.16
-    - dqsegdb2==1.0.1
-    - dynesty==1.0.1
-    - emcee==3.0.2
-    - entrypoints==0.3
-    - future==0.18.2
-    - gwdatafind==1.0.4
-    - gwosc==0.5.6
-    - gwpy==2.1.1
-    - h5py==3.1.0
-    - idna==2.10
-    - imagesize==1.2.0
-    - importlib-metadata==3.7.0
-    - ipython-genutils==0.2.0
-    - jinja2==2.11.3
-    - jsonschema==3.2.0
-    - jupyter-client==6.1.11
-    - jupyter-core==4.7.1
-    - jupyterlab-pygments==0.1.2
-    - kiwisolver==1.3.1
-    - lalsuite==7.3
-    - ligo-gracedb==2.7.6
-    - ligo-segments==1.3.0
-    - ligotimegps==2.0.1
-    - llvmlite==0.35.0
[…]
