### 2022-06-24 b7d8260
docs-environment.yml: move lintegrate into conda install section
