### 2022-06-29 daa4eff
heterodyne.py: use scipy filter functions for creating and performing filtering
