+++ cwinpy/heterodyne/base.py
@@ -173,7 +173,7 @@ class Heterodyne(object):
     crop: int
         The number of seconds to crop from the start and end of data segments
         to remove filter impulse effects and issues prior to lock-loss.
-        Default is 60 seconds.
+        Default is 600 seconds.
     includessb: bool
         Set this to True to include removing the modulation of the signal due
         to Solar System motion and relativistic effects (e.g., Roemer,
@@ -262,7 +262,7 @@ class Heterodyne(object):
         filterknee=0.5,
         resamplerate=1.0,
         freqfactor=2,
-        crop=60,
+        crop=600,
         includessb=False,
         includebsb=False,
         includeglitch=False,
@@ -2087,15 +2087,7 @@ class Heterodyne(object):
                 het.cwinpy_heterodyne_pipeline_config = cf
 
             # save filter history from the forward pass
-            history = []
-            for idx in range(len(self._filters[pulsar])):
-                history.append(
-                    (
-                        self._filters[pulsar][idx][0].history.data,
-                        self._filters[pulsar][idx][1].history.data,
-                    )
-                )
-            het.filter_history = history
+            het.filter_history = self._filter_history[pulsar]
 
             het.write(self.outputfiles[pulsar], overwrite=True)
 
@@ -2258,11 +2250,10 @@ class Heterodyne(object):
 
         return hetfiles, isfile
 
-    def _setup_filters(self, filterknee, samplerate):
+    def _setup_filters(self, filterknee, samplerate, n=9):
         """
-        Set up the 9th order low-pass Butterworth filters to apply to the data
-        for each pulsars. This is performed by applying a 3rd order filter
-        three times.
+        Set up the nth order low-pass Butterworth filter to apply to the data
+        for each pulsars.
 
         Parameters
         ----------
@@ -2270,38 +2261,26 @@ class Heterodyne(object):
             The filter knee frequency in Hz.
         samplerate: float
             The data sample rate in Hz
+        n: int
+            The filter order
         """
+
+        from scipy.signal import butter
 
         if not hasattr(self, "_filters"):
             self._filters = {}
         else:
             return
 
-        wc = np.tan(np.pi * filterknee / samplerate)
-
-        # set zero pole gain values
-        zpg = lal.CreateCOMPLEX16ZPGFilter(0, 3)
-        zpg.poles.data[0] = (wc * np.sqrt(3.0) / 2.0) + 1j * (wc * 0.5)
-        zpg.poles.data[1] = 1j * wc
-        zpg.poles.data[2] = -(wc * np.sqrt(3.0) / 2.0) + 1j * (wc * 0.5)
-        zpg.gain = 1j * wc * wc * wc
-        lal.WToZCOMPLEX16ZPGFilter(zpg)
+        self._filter_history = {}
 
         for pulsar in self.pulsars:
-            self._filters[pulsar] = []
+            self._filter_history[pulsar] = None
 
-            # create IIR filters
-            for i in range(3):
-                filterRe = lal.CreateREAL8IIRFilter(zpg)
-                filterIm = lal.CreateREAL8IIRFilter(zpg)
-
-                if hasattr(self, "_filter_history"):
-                    if pulsar in self._filter_history:
-                        # set previous history
-                        filterRe.history.data = self._filter_history[pulsar][i][0]
-                        filterIm.history.data = self._filter_history[pulsar][i][1]
-
-                self._filters[pulsar].append((filterRe, filterIm))
+            # create filters in sos format for numerical stability
+            filterRe = butter(n, filterknee, fs=samplerate, btype="low", output="sos")
[…]
