### 2022-06-29 0b07888
Set initial filter value to better value which avoids the huge impulse response - this means that the default cropping can be put back to 60 seconds
