### 2022-06-30 bd51072
base.py: remove forwardsonly argument from _filter_data as this is never exposed elsewhere in the code - the function will now always run the filter in forwards and reverse
