### 2022-06-30 d6535bf
Expose the filter order so that it can be set by the user. The default will be 9.
