+++ docs/heterodyne/cwinpy_heterodyne_pipeline.ini
@@ -253,6 +253,10 @@ excludeflags = {"H1": ["H1:DCS-BAD_STRAIN_H1_HOFT_C02:1,H1:DCS-BAD_KAPPA_BASIC_C
 # stage and 0.5 Hz if heterodyning over two stages.
 ;filterknee =
 
+# The order of the low-pass Butterworth filter applied after heterodyning the
+# data. The default is 9.
+;filterorder =
+
 # The number of seconds to crop from the start and end of data segments after
 # the initial heterodyned to remove filter impulse effects and issues prior to
 # lock-loss. Default is 60 seconds.
