### 2022-07-05 a112c42
data.py: minor fix to plotting with outlier removal
