### 2022-07-06 42de26e
Fix the setting of the Bayesian Blocks mininum chunks length
