### 2022-07-06 #4476 @mattpitkin
@dfm Thanks. Another potential reviewer in additional to those [listed above](https://github.com/openjournals/joss-reviews/issues/4476#issuecomment-1156238095) is jcbayley. I'm not sure if this would be perceived as a conflict of interest, but jcbayley has been involved in reviewing the code and paper internally for the LIGO-Virgo-KAGRA collaborations, so would already have a head start.
