### 2022-07-07 a4b21f9
plot.py: allow data paths as pathlib.Path objects
