### 2022-07-07 25e9e02
readers.py: fix black formatting failure
