### 2022-07-12 2e887ed
Allow consistent pulsarfiles option in cwinpy_pe_pipeline configuration file
