### 2022-07-14 e2acdef
Fix UpperlimitTable plot when using non-Quantity value column
