+++ cwinpy/pe/peutils.py
@@ -1340,14 +1340,17 @@ class UpperLimitTable(QTable):
             else:
                 colname = acolumn
 
-            axisdata[xy]["data"] = self[colname].value.copy()
+            if hasattr(self[colname], "value"):
+                axisdata[xy]["data"] = self[colname].value.copy()
+            else:
+                axisdata[xy]["data"] = self[colname].copy()
             axisdata[xy]["label"] = colname
 
         # scale rotation frequency to GW frequency if necessary
         if isinstance(column, str) and axisdata["y"]["label"].upper().startswith(
             ("H0", "ELL", "Q22", "C22")
         ):
-            axisdata["x"]["data"] *= 2.0  # GWs ate twice frot
+            axisdata["x"]["data"] *= 2.0  # GWs at twice frot
             axisdata["x"]["label"] = "F0GW"
 
         # start creating figure
