### 2022-07-14 d961e67
Add autoscaling to show spin-down limits
