### 2022-07-15 #4568 @editorialbot
[REVIEW]: CWInPy: a Python package for inference with continuous gravitational-wave signals from pulsars
**Submitting author:** @mattpitkin (<a href="http://orcid.org/0000-0003-4548-526X">Matthew Pitkin</a>)
**Repository:** https://github.com/cwinpy/cwinpy
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.9.0
