### 2022-07-15 #4568 @mattpitkin
@gregoryashton @colmtalbot thanks very much for agreeing to perform the review. When looking at the documentation it's worth looking at the [latest version](https://cwinpy.readthedocs.io/en/latest/index.html) rather than the stable version. The [internal LVK review](https://git.ligo.org/groups/cwinpy/-/epics/1) (you should be able to see that material if you want to) of the code should be concluding soon, and I expect to release a v1.0.0 version at that point. If you have any issues feel free to submit them [here](https://git.ligo.org/groups/cwinpy/-/issues).
