+++ paper/paper.bib
@@ -499,3 +499,18 @@
   adsnote       = {Provided by the SAO/NASA Astrophysics Data System},
   doi           = {10.48550/arXiv.2111.13106}
 }
+
+@article{pesummary,
+    author = {{Hoy}, C. and {Raymond}, V.},
+    title = "{PESummary: the code agnostic Parameter Estimation Summary page builder}",
+    eprint = "2006.06639",
+    archivePrefix = "arXiv",
+    primaryClass = "astro-ph.IM",
+    reportNumber = "LIGO-P2000156",
+    doi = "10.1016/j.softx.2021.100765",
+    journal = "SoftwareX",
+    volume = "15",
+    pages = "100765",
+    year = "2021"
+}
+
