+++ setup.cfg
@@ -31,6 +31,7 @@ setup_requires =
     wheel
 install_requires =
     appdirs
+    arby
     astropy
     bilby >= 1.1.4
     configargparse
