### 2022-07-18 f017759
roq.py: make class able to use a more generic model
