+++ cwinpy/pe/roq.py
@@ -27,10 +27,23 @@ class GenerateROQ:
         self.ntraining = ntraining
 
         # generate training data
-        self.generate_training_set()
+        ts = self.generate_training_set()
+
+        if self.kwargs.get("store_training_data", False):
+            # store training data if needed
+            self.training_data = ts
 
         # generate the reduced basis
-        self.generate_reduced_basis()
+        self.generate_reduced_basis(ts)
+
+        # likelihood term for squared data
+        sigma = self.kwargs.get("sigma", 1.0)
+        if self.data.dtype == complex:
+            self._K = np.sum((self.data.real / sigma) ** 2) + np.sum(
+                (self.data.imag / sigma) ** 2
+            )
+        else:
+            self._K = np.sum((self.data / sigma) ** 2)
 
     @property
     def x(self):
@@ -84,14 +97,12 @@ class GenerateROQ:
 
         self._priors = priors
 
-    def generate_training_set(self):
-        """
-        Generate ``n`` model instances, using values drawn from the prior.
-        """
+    @property
+    def model(self):
+        return self._model
 
-        self._training_set_model = []
-
-        model = self.kwargs.get("model", HeterodynedCWSimulator)
+    @model.setter
+    def model(self, model):
         if isinstance(model, HeterodynedCWSimulator):
             if "par" not in self.kwargs:
                 raise ValueError("'par', parameter file must be in keyword arguments")
@@ -105,12 +116,48 @@ class GenerateROQ:
                 par, det, self.x, usetempo2=self.kwargs.get("usetempo2", False)
             )
 
-            modelfunc = model.model
+            self._initial_par = deepcopy(par)
+            self._par = deepcopy(par)
 
-            # copy of heterodyned parameters
-            newpar = deepcopy(self.par)
+            self._model = model.model
         else:
-            modelfunc = model
+            self._model = model
+
+    def _model_at_nodes(self, xr, xi, x2):
+        if hasattr(self, "_initial_par"):
+            det = self.kwargs.get("det")
+            modelr = HeterodynedCWSimulator(
+                self._initial_par,
+                det,
+                xr,
+                usetempo2=self.kwargs.get("usetempo2", False),
+            )
+            self._interpolated_model_real = modelr.model
+            modeli = HeterodynedCWSimulator(
+                self._initial_par,
+                det,
+                xi,
+                usetempo2=self.kwargs.get("usetempo2", False),
+            )
+            self._interpolated_model_imag = modeli.model
+            model2 = HeterodynedCWSimulator(
+                self._initial_par,
+                det,
+                x2,
+                usetempo2=self.kwargs.get("usetempo2", False),
+            )
+            self._interpolated_model2 = model2.model
+        else:
+            self._interpolated_model = self.model
+
+    def generate_training_set(self):
+        """
+        Generate ``n`` model instances, using values drawn from the prior.
+        """
+
+        training_set_model = []
+
+        self.model = self.kwargs.get("model", HeterodynedCWSimulator)
 
         # draw values from prior
         samples = self.priors.sample(self.ntraining)
@@ -122,12 +169,12 @@ class GenerateROQ:
             for prior in samples:
                 self.samples[-1].append(samples[prior][i])
 
-                if isinstance(model, HeterodynedCWSimulator):
-                    newpar[prior] = samples[prior][i]
+                if hasattr(self, "_par"):
+                    self._par[prior] = samples[prior][i]
 
-            if isinstance(model, HeterodynedCWSimulator):
-                m = modelfunc(
-                    deepcopy(newpar),
+            if hasattr(self, "_par"):
+                m = self.model(
+                    deepcopy(self._par),
                     outputampcoeffs=False,
[…]
