### 2022-07-20 98a7cfa
Complete GenerateROQ class
