+++ cwinpy/pe/roq.py
@@ -1,30 +1,73 @@
 from copy import deepcopy
 
+import lal
 import numpy as np
 from arby import reduced_basis
 from bilby.core.prior import PriorDict
+from numba import jit
 
 from ..signal import HeterodynedCWSimulator
+from ..utils import logfactorial
 
 
 class GenerateROQ:
-    def __init__(self, data, x, priors, ntraining=1000, **kwargs):
+    def __init__(self, data, x, priors, **kwargs):
         """
         Generate the Reduced Order Quadrature for the parameters in the prior.
 
         Parameters
         ----------
         data: array_like
-            The data
+            The data set (real or complex) for which the ROQ will be produced.
+        x: array_like
+            The values at which the data is defined (e.g., a set of times).
+        priors: PriorDict
+            A :class:`~bilby.core.prior.PriorDict` containing the parameters
+            and prior distributions, which will be used to generate the model
+            reduced basis set.
+        ntraining: int
+            The number of training example models to be generated from which to
+            produce the reduced basis set. Defaults to 5000.
+        sigma: array_like, float
+            The noise standard deviation for the data, which is assumed to be
+            stationary over the data. If not given, this will be assume to be
+            1.0 for all data points.
+        store_training_data: bool
+            If set to True the class will store the array of models used to
+            generate the reduced basis. Note that if using this class multiple
+            times, storing the data could become memory intensive. Defaults to
+            False.
+        greedy_tol: float
+            The tolerance to use when generating the reduced basis. Defaults to
+            1e-12.
+        model: func
+            The model function used to generate the training examples. By
+            default this assumes the
+            :class:`cwinpy.signal.HeterodynedCWSimulator` as the model (see
+            Other Parameters below), but it can be an arbitrary Python function
+            provided the function has an initial positional argument that takes
+            in the ``x`` values, followed by keyword arguments for the other
+            function parameters as defined in the priors.
+
+        Other Parameters
+        ----------------
+        par: PulsarParameters
+            A :class:`cwinpy.parfile.PulsarParameters` or parameter file
+            containing the parameters used for heterodyning the data.
+        det: str
+            The name of the detector for which the signal will be modelled.
+        usetempo2: bool
+            Set to True if the signal model phase evolution should be
+            calculated using Tempo2. Defaults to False.
         """
 
+        self.kwargs = kwargs
         self.x = x
         self.data = data
         self.priors = priors
-        self.kwargs = kwargs
 
         # the number of training examples to use
-        self.ntraining = ntraining
+        self.ntraining = self.kwargs.get("ntraining", 5000)
 
         # generate training data
         ts = self.generate_training_set()
@@ -35,15 +78,6 @@ class GenerateROQ:
 
         # generate the reduced basis
         self.generate_reduced_basis(ts)
-
-        # likelihood term for squared data
-        sigma = self.kwargs.get("sigma", 1.0)
-        if self.data.dtype == complex:
-            self._K = np.sum((self.data.real / sigma) ** 2) + np.sum(
-                (self.data.imag / sigma) ** 2
-            )
-        else:
-            self._K = np.sum((self.data / sigma) ** 2)
 
     @property
     def x(self):
@@ -86,6 +120,17 @@ class GenerateROQ:
 
         self._data = dataarr
 
+        # likelihood term for squared data
+        self._sigma = self.kwargs.get("sigma", 1.0)
+
+        if not isinstance(self._sigma, (int, float)):
[…]
