+++ cwinpy/pe/likelihood.py
@@ -794,7 +794,7 @@ class TargetedPulsarLikelihood(bilby.core.likelihood.Likelihood):
             # compute "Chi-squared"
             chisquare = productsreal["ddotd"][i] - 2.0 * sumdatamodel + summodel
 
-            print(sumdatamodel, summodel, productsreal["ddotd"][i])
+            # print(sumdatamodel, summodel, productsreal["ddotd"][i])
 
             if likelihood == "gaussian":
                 loglikelihood += 0.5 * chisquare
