### 2022-07-25 #4568 @GregoryAshton
@mattpitkin regarding the paper, there is a typo in the summary: a missing "to" between "user" and "search" in paragraph 2.

I also wondered if there is an appropriate link to documentation on what an "ephemeris" file should be? While this is a well-understood concept in radio pulsar astronomy, users of this package may be unfamiliar and unable to create an ephemeris. Moreover, the definition of an ephemeris may yet change (there are standard formats I think). Would you consider adding to the documentation the fields that cwinpy expects (e.g. key/value information) and perhaps linking that from the paper?
