### 2022-07-25 fac37e7
roq.py: split real and imaginary ROM generation
