+++ cwinpy/pe/roq.py
@@ -129,6 +129,8 @@ class GenerateROQ:
             if self._sigma <= 0.0:
                 raise ValueError("sigma must be a positive number")
 
+        self.is_complex = True if self._data.dtype == complex else False
+
         self._K = np.vdot(self._data, self._data).real
 
     @property
@@ -242,12 +244,50 @@ class GenerateROQ:
         """
 
         # reduced basis for the model training set
-        rb_model = reduced_basis(
-            training_set=np.array(training),
-            physical_points=self.x,
-            greedy_tol=self.kwargs.get("greedy_tol", 1e-12),
-            normalize=True,
-        )
+        if self.is_complex:
+            rb_model_real = reduced_basis(
+                training_set=np.array(training).real,
+                physical_points=self.x,
+                greedy_tol=self.kwargs.get("greedy_tol", 1e-12),
+                normalize=True,
+            )
+            rb_model_imag = reduced_basis(
+                training_set=np.array(training).imag,
+                physical_points=self.x,
+                greedy_tol=self.kwargs.get("greedy_tol", 1e-12),
+                normalize=True,
+            )
+
+            self._Dvec_real = np.einsum(
+                "i,ji->j", self.data.real, rb_model_real.basis.data
+            )
+            self._Dvec_imag = np.einsum(
+                "i,ji->j", self.data.imag, rb_model_imag.basis.data
+            )
+
+            nodes_real = np.array(sorted(rb_model_real.basis.eim_.nodes))
+            nodes_imag = np.array(sorted(rb_model_imag.basis.eim_.nodes))
+
+            self._node_indices = np.union1d(nodes_real, nodes_imag)
+            self._Bmat_real = np.array(
+                [row[nodes_real] for row in rb_model_real.basis.data]
+            )
+            self._Bmat_imag = np.array(
+                [row[nodes_imag] for row in rb_model_imag.basis.data]
+            )
+        else:
+            rb_model = reduced_basis(
+                training_set=np.array(training),
+                physical_points=self.x,
+                greedy_tol=self.kwargs.get("greedy_tol", 1e-12),
+                normalize=True,
+            )
+
+            self._Dvec = np.einsum("i,ji->j", self.data, rb_model.basis.data)
+
+            nodes = np.array(sorted(rb_model.basis.eim_.nodes))
+            self._node_indices = nodes
+            self._Bmat = np.array([row[nodes] for row in rb_model.basis.data])
 
         # reduced basis for the squared model training set
         rb_model_squared = reduced_basis(
@@ -257,32 +297,26 @@ class GenerateROQ:
             normalize=True,
         )
 
-        # get pre-summed elements
-        self._Dvec = np.einsum("i,ji->j", self.data, rb_model.basis.data)
         self._Bvec = np.einsum("ji->j", rb_model_squared.basis.data)
-
-        self._Bmat = np.array(
-            [row[rb_model.basis.eim_.nodes] for row in rb_model.basis.data]
-        )
-        self._B2mat = np.array(
-            [
-                row[rb_model_squared.basis.eim_.nodes]
-                for row in rb_model_squared.basis.data
-            ]
-        )
-
-        nodes = rb_model.basis.eim_.nodes
-        nodes2 = rb_model_squared.basis.eim_.nodes
-
-        self.nbases = len(nodes)
-        self.nbases2 = len(nodes2)
+        nodes2 = np.array(sorted(rb_model_squared.basis.eim_.nodes))
+        self._B2mat = np.array([row[nodes2] for row in rb_model_squared.basis.data])
 
         # setup model function at the interpolation nodes
-        self._node_indices = np.union1d(nodes, nodes2)
+        self._node_indices = np.union1d(self._node_indices, nodes2)
 
         self._x_nodes = self.x[self._node_indices]
[…]
