### 2022-07-26 #4568 @mattpitkin
@GregoryAshton Thanks for spotting the typo. I'll get that fixed.

I think the appropriate reference for the ephemeris file format is the Tempo2 [manual](https://www.jb.man.ac.uk/~pulsar/Resources/tempo2_manual.pdf). I'll add a link to that. I can also add to the documentation explicit information about the required parameters that must be in a par file for CWInPy.
