### 2022-07-26 4e40699
Allow ET to be specified as a detector for noise generation
