### 2022-07-26 07d3fc0
Fix typo in summary - refs https://github.com/openjournals/joss-reviews/issues/4568#issuecomment-1193897450
