### 2022-07-26 55b05f8
Add info on parameter file specification - refs https://github.com/openjournals/joss-reviews/issues/4568#issuecomment-1193897450
