### 2022-07-26 008bf76
Add info on parameter file specification - refs https://github.com/openjournals/joss-reviews/issues/4568#issuecomment-1193897450
