### 2022-07-26 3e81820
Add info to paper on parameter file specification - refs https://github.com/openjournals/joss-reviews/issues/4568#issuecomment-1193897450
