+++ paper/paper.bib
@@ -130,6 +130,12 @@
   year         = {2018}
 }
 
+@misc{tempo2manual,
+  author       = {{Hobbs}, G. and {Edwards}, R.},
+  title        = {{TEMPO2} user manual},
+  url          = {https://www.jb.man.ac.uk/~pulsar/Resources/tempo2_manual.pdf},
+}
+
 @article{WETTE2020100634,
   title    = {{SWIGLAL: Python and Octave interfaces to the LALSuite gravitational-wave data analysis libraries}},
   journal  = {{SoftwareX}},
