+++ paper/paper.bib
@@ -134,6 +134,8 @@
   author       = {{Hobbs}, G. and {Edwards}, R.},
   title        = {{TEMPO2} user manual},
   url          = {https://www.jb.man.ac.uk/~pulsar/Resources/tempo2_manual.pdf},
+  year         = {2006},
+  note         = {version 2.0},
 }
 
 @article{WETTE2020100634,
