### 2022-07-26 #4568 @mattpitkin
@GregoryAshton I've added a section to the documentation about the parameter file specification https://cwinpy.readthedocs.io/en/latest/pipelines.html#source-parameter-specification. I've also linked to this in the paper and added a reference to the Tempo2 manual.
