### 2022-07-26 #4568 @ColmTalbot
@mattpitkin I've started the review and opened an [issue](https://git.ligo.org/cwinpy/cwinpy/-/issues/84) on the development repo with some installation/testing issues.

I'm still churning through a few of the pipeline examples to get a better feel for the repo as a whole.

I have a few comments that came up when reading through the paper/docs.

- Citations: a review or textbook citation for the first sentence (about CW sources) may help non-specialist readers, consider citing the OSG (https://osg-htc.org/acknowledging.html).
- It's unclear to me which of the heterodyning methods are preferred (if any) or when they should be used.
