### 2022-07-27 440f168
Add review article reference for first sentence and add OSG reference  - refs https://github.com/openjournals/joss-reviews/issues/4568#issuecomment-1195947115
