+++ paper/paper.bib
@@ -522,3 +522,32 @@
     year = "2021"
 }
 
+@article{Pordes_2007,
+	doi = {10.1088/1742-6596/78/1/012057},
+	url = {https://doi.org/10.1088/1742-6596/78/1/012057},
+	year = 2007,
+	month = {jul},
+	publisher = {{IOP} Publishing},
+	volume = {78},
+	pages = {012057},
+	author = {Ruth Pordes and Don Petravick and Bill Kramer and Doug Olson and Miron Livny and Alain Roy and Paul Avery and Kent Blackburn and Torre Wenaus and Frank Würthwein and Ian Foster and Rob Gardner and Mike Wilde and Alan Blatecky and John McGee and Rob Quick},
+	title = {The open science grid},
+}
+
+@ARTICLE{2022arXiv220606447R,
+       author = {{Riles}, Keith},
+        title = "{Searches for Continuous-Wave Gravitational Radiation}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - High Energy Astrophysical Phenomena, General Relativity and Quantum Cosmology},
+         year = 2022,
+        month = jun,
+          eid = {arXiv:2206.06447},
+        pages = {arXiv:2206.06447},
+archivePrefix = {arXiv},
+       eprint = {2206.06447},
+ primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022arXiv220606447R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+          doi = {10.48550/arXiv.2206.06447}
+}
+
