### 2022-07-27 #4568 @mattpitkin
@ColmTalbot Thanks for reporting the issue. I'll get those fixed.

> Citations: a review or textbook citation for the first sentence (about CW sources) may help non-specialist readers, consider citing the OSG (https://osg-htc.org/acknowledging.html).

I've added a reference to a review article and also cited the OSG paper.

> It's unclear to me which of the heterodyning methods are preferred (if any) or when they should be used.

I assume you mean `cwinpy_heterodyne` versus `cwinpy_heterodyne_pipeline`? Basically, the latter should be used for almost all practical purposes, as it will generate and run the heterodyne as a Condor DAG. The former is really just for quick testing on short bits of data. I'll try and make that clear.

> I would quibble with the phrasing "the Bayesian evidence for the signal model given the data". To me, the evidence is the probability of the data given the model.

Yep, that was a mistake on my part. I've fixed it now.
