### 2022-07-27 ca3c6ab
Address further comments from https://github.com/openjournals/joss-reviews/issues/4568#issuecomment-1196575214
