### 2022-07-27 02b5822
Exclude paper branch from CI
