+++ paper/paper.bib
@@ -530,6 +530,7 @@
 	publisher = {{IOP} Publishing},
 	volume = {78},
 	pages = {012057},
+  journal = {Journal of Physics: Conference Series},
 	author = {Ruth Pordes and Don Petravick and Bill Kramer and Doug Olson and Miron Livny and Alain Roy and Paul Avery and Kent Blackburn and Torre Wenaus and Frank Würthwein and Ian Foster and Rob Gardner and Mike Wilde and Alan Blatecky and John McGee and Rob Quick},
 	title = {The open science grid},
 }
