### 2022-07-27 818a100
roq.py: calculate LU decomposition and use scipys lu_solve for speed improvement
