+++ cwinpy/pe/roq.py
@@ -4,7 +4,7 @@ import lal
 import numpy as np
 from arby import reduced_basis
 from bilby.core.prior import PriorDict
-from numba import jit
+from scipy.linalg import lu_factor, lu_solve
 
 from ..signal import HeterodynedCWSimulator
 from ..utils import logfactorial
@@ -249,13 +249,11 @@ class GenerateROQ:
                 training_set=np.array(training).real,
                 physical_points=self.x,
                 greedy_tol=self.kwargs.get("greedy_tol", 1e-12),
-                normalize=True,
             )
             rb_model_imag = reduced_basis(
                 training_set=np.array(training).imag,
                 physical_points=self.x,
                 greedy_tol=self.kwargs.get("greedy_tol", 1e-12),
-                normalize=True,
             )
 
             self._Dvec_real = np.einsum(
@@ -269,37 +267,39 @@ class GenerateROQ:
             nodes_imag = np.array(sorted(rb_model_imag.basis.eim_.nodes))
 
             self._node_indices = np.union1d(nodes_real, nodes_imag)
-            self._Bmat_real = np.array(
-                [row[nodes_real] for row in rb_model_real.basis.data]
+            self._Bmat_lu_real = lu_factor(
+                np.array([row[nodes_real] for row in rb_model_real.basis.data]).T
             )
-            self._Bmat_imag = np.array(
-                [row[nodes_imag] for row in rb_model_imag.basis.data]
+            self._Bmat_lu_imag = lu_factor(
+                np.array([row[nodes_imag] for row in rb_model_imag.basis.data]).T
             )
         else:
             rb_model = reduced_basis(
                 training_set=np.array(training),
                 physical_points=self.x,
                 greedy_tol=self.kwargs.get("greedy_tol", 1e-12),
-                normalize=True,
             )
 
             self._Dvec = np.einsum("i,ji->j", self.data, rb_model.basis.data)
 
             nodes = np.array(sorted(rb_model.basis.eim_.nodes))
             self._node_indices = nodes
-            self._Bmat = np.array([row[nodes] for row in rb_model.basis.data])
+            self._Bmat_lu = lu_factor(
+                np.array([row[nodes] for row in rb_model.basis.data]).T
+            )
 
         # reduced basis for the squared model training set
         rb_model_squared = reduced_basis(
             training_set=np.array((training * np.conj(training)).real),
             physical_points=self.x,
             greedy_tol=self.kwargs.get("greedy_tol", 1e-12),
-            normalize=True,
         )
 
         self._Bvec = np.einsum("ji->j", rb_model_squared.basis.data)
         nodes2 = np.array(sorted(rb_model_squared.basis.eim_.nodes))
-        self._B2mat = np.array([row[nodes2] for row in rb_model_squared.basis.data])
+        self._B2mat_lu = lu_factor(
+            np.array([row[nodes2] for row in rb_model_squared.basis.data]).T
+        )
 
         # setup model function at the interpolation nodes
         self._node_indices = np.union1d(self._node_indices, nodes2)
@@ -331,9 +331,6 @@ class GenerateROQ:
             likelihood ``"gaussian"`` or Student's t-likelihood ``"studentst"``.
             As with the :class:`cwinpy.pe.likelihood.TargetedPulsarLikelihood`,
             the default is Student's t.
-        numba: bool
-            This defaults to True in which case the likelihood defined using
-            numba will be used.
         """
 
         likelihood = kwargs.get("likelihood", "studentst").lower()
@@ -358,100 +355,34 @@ class GenerateROQ:
         # generate the model at the interpolation nodes
         model = self._model_short(*pos, **self.__model_kwargs)
 
-        if kwargs.get("numba", True):
-            if self.is_complex:
[…]
