### 2022-07-27 #4568 @ColmTalbot
> I assume you mean cwinpy_heterodyne versus cwinpy_heterodyne_pipeline? Basically, the latter should be used for almost all practical purposes, as it will generate and run the heterodyne as a Condor DAG. The former is really just for quick testing on short bits of data. I'll try and make that clear.

Sorry, I was unclear, I meant LAL vs Tempo.
