+++ examples/example_targeted_pulsar.py
@@ -7,7 +7,8 @@ pulsar-type search.
 
 import numpy as np
 from bilby.core.prior import PriorDict, Uniform
-from cwinpy import HeterodynedData, TargetedPulsarLikelihood
+from cwinpy import HeterodynedData
+from cwinpy.pe.likelihood import TargetedPulsarLikelihood
 from matplotlib import pyplot as pl
 
 # create a fake pulsar parameter file
