### 2022-07-27 d878d34
Use --accounting-group rather than --accounting-group-tag for pipeline CLIs to be consistent with config files
