+++ cwinpy/heterodyne/heterodyne.py
@@ -1755,7 +1755,7 @@ def heterodyne_pipeline(**kwargs):
         The length of data (in seconds) into which to split the individual
         analysis jobs. By default this is set to 86400, i.e., one day. If this
         is set to 0, then the whole dataset is treated as a single job.
-    accounting_group_tag: str
+    accounting_group: str
         For LVK users this sets the computing accounting group tag.
     usetempo2: bool
         Set this flag to use Tempo2 (if installed) for calculating the signal
@@ -1871,7 +1871,7 @@ def heterodyne_pipeline(**kwargs):
             ),
         )
         optional.add_argument(
-            "--accounting-group-tag",
+            "--accounting-group",
             dest="accgroup",
             help=("For LVK users this sets the computing accounting group tag"),
         )
@@ -2007,7 +2007,7 @@ def heterodyne_quick_setup(args, **kwargs):
 
     if args.accgroup is not None:
         configfile["heterodyne_job"]["accounting_group"] = kwargs.get(
-            "accounting_group_tag", args.accgroup
+            "accounting_group", args.accgroup
         )
 
     # add pulsars/pulsar ephemerides
