### 2022-07-28 #4568 @GregoryAshton
@mattpitkin I tried installing from source. using `pip install .` First, it may be worth noting you need a `gcc` compiler. Second, I then got stuck with `lintegrate`:

```console
  Building wheel for lintegrate (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Building wheel for lintegrate (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [19 lines of output]
      running bdist_wheel
      running build
      running build_py
      creating build
      creating build/lib.linux-x86_64-cpython-39
      creating build/lib.linux-x86_64-cpython-39/lintegrate
      copying lintegrate/__init__.py -> build/lib.linux-x86_64-cpython-39/lintegrate
      copying lintegrate/lintegrate.pxd -> build/lib.linux-x86_64-cpython-39/lintegrate
      running build_ext
      building 'lintegrate.lintegrate' extension
      creating build/temp.linux-x86_64-cpython-39
      creating build/temp.linux-x86_64-cpython-39/lintegrate
      creating build/temp.linux-x86_64-cpython-39/src
      gcc -pthread -B /home/greg/anaconda3/envs/cw/compiler_compat -Wno-unused-result -Wsign-compare -DNDEBUG -O2 -Wall -fPIC -O2 -isystem /home/greg/anaconda3/envs/cw/include -I/home/greg/anaconda3/envs/cw/include -fPIC -O2 -isystem /home/greg/anaconda3/envs/cw/include -fPIC -I/tmp/pip-build-env-bjexc8c2/overlay/lib/python3.9/site-packages/numpy/core/include -I/home/greg/anaconda3/envs/cw/include -Isrc -Ilintegrate -I/home/greg/anaconda3/envs/cw/include/python3.9 -c lintegrate/lintegrate.c -o build/temp.linux-x86_64-cpython-39/lintegrate/lintegrate.o -Wall -DHAVE_PYTHON_LINTEGRATE -O3 -Wextra -m64 -ffast-math -fno-finite-math-only -march=native -funroll-loops
      cc1: error: bad value (‘tigerlake’) for ‘-march=’ switch
      cc1: note: valid arguments to ‘-march=’ switch are: nocona core2 nehalem corei7 westmere sandybridge corei7-avx ivybridge core-avx-i haswell core-avx2 broadwell skylake skylake-avx512 cannonlake icelake-client icelake-server cascadelake bonnell atom silvermont slm goldmont goldmont-plus tremont knl knm x86-64 eden-x2 nano nano-1000 nano-2000 nano-3000 nano-x2 eden-x4 nano-x4 k8 k8-sse3 opteron opteron-sse3 athlon64 athlon64-sse3 athlon-fx amdfam10 barcelona bdver1 bdver2 bdver3 bdver4 znver1 znver2 btver1 btver2 native
      cc1: error: bad value (‘tigerlake’) for ‘-mtune=’ switch
      cc1: note: valid arguments to ‘-mtune=’ switch are: nocona core2 nehalem corei7 westmere sandybridge corei7-avx ivybridge core-avx-i haswell core-avx2 broadwell skylake skylake-avx512 cannonlake icelake-client icelake-server cascadelake bonnell atom silvermont slm goldmont goldmont-plus tremont knl knm intel x86-64 eden-x2 nano nano-1000 nano-2000 nano-3000 nano-x2 eden-x4 nano-x4 k8 k8-sse3 opteron opteron-sse3 athlon64 athlon64-sse3 athlon-fx amdfam10 barcelona bdver1 bdver2 bdver3 bdver4 znver1 znver2 btver1 btver2 generic native
      error: command '/usr/bin/gcc' failed with exit code 1
      [end of output]

  ERROR: Failed building wheel for lintegrate
  Building wheel for ads (setup.py) ... done
  Created wheel for ads: filename=ads-0.12.3-py3-none-any.whl size=37390 sha256=ce18a59f07b88627c05e42941794077a040cec4b715f4869a67fcd7b66aede05
  Stored in directory: /home/greg/.cache/pip/wheels/c8/27/d9/ce264eacc76ed8c985c02c9d1a02e962493d9c5583ec5d947e
  Building wheel for httpretty (setup.py) ... done
  Created wheel for httpretty: filename=httpretty-0.8.10-py3-none-any.whl size=24014 sha256=7317ebe86932ee85b8b22dfeedb823580245de1496096a5b6ae784b8e308f5f5
  Stored in directory: /home/greg/.cache/pip/wheels/7c/1c/cf/2d9a759c4854ba9064ef81c6c858329f4f2ebcc1739c5eeab9
Successfully built cwinpy ads httpretty
Failed to build lintegrate
ERROR: Could not build wheels for lintegrate, which is required to install pyproject.toml-based projects
```

`lintegrate` can be installed from conda, but it seems to fail when installing with pip.

This was using `python` version 3.9.12. I'm not sure this is a show stopper, but on the other hand it is unclear how to solve.
