### 2022-07-28 #4568 @GregoryAshton
@mattpitkin, when running the make fake data example, I get hit again by more gsl issues

```console
Traceback (most recent call last):
  File "/home/greg/test/cwinpy/mfd.py", line 6, in <module>
    from cwinpy import PulsarParameters
  File "/home/greg/anaconda3/envs/cw/lib/python3.9/site-packages/cwinpy/__init__.py", line 4, in <module>
    from . import iostream
  File "/home/greg/anaconda3/envs/cw/lib/python3.9/site-packages/cwinpy/iostream/__init__.py", line 1, in <module>
    from . import ascii  # pylint: disable=unused-import
  File "/home/greg/anaconda3/envs/cw/lib/python3.9/site-packages/cwinpy/iostream/ascii.py", line 5, in <module>
    from ..data import HeterodynedData
  File "/home/greg/anaconda3/envs/cw/lib/python3.9/site-packages/cwinpy/data.py", line 11, in <module>
    import lal
  File "/home/greg/anaconda3/envs/cw/lib/python3.9/site-packages/lal/__init__.py", line 2, in <module>
    from .lal import *
  File "/home/greg/anaconda3/envs/cw/lib/python3.9/site-packages/lal/lal.py", line 13, in <module>
    from ._lal import *
ImportError: libgsl.so.25: cannot open shared object file: No such file or directory
```

I'm fairly confused because I have a gsl compiler and that standard build-essentials packages.
