### 2022-07-28 #4568 @mattpitkin
These gsl and lintegrate issues have be my bane in packaging this! Whenever I think it works ok, a case turns up where it doesn't. I'm in holiday until 7th August, but will take a look after that. I know he's currently on holiday too, but I'll cc @duncanmmacleod in case he has any insight.
