### 2022-07-28 #4568 @mattpitkin
It looks like I need to remove the `march=native` from the lintegrate
compilation command in its setup.py. I'll try this and release a new
version.
