### 2022-07-28 #4568 @mattpitkin
There's now a new version of lintegrate (v0.1.14) available on PyPI that has the `march=native` flag removed, and should hopefully compile on tigerly
