### 2022-08-08 29f7c9a
setup.cfg: use v7.6 of lalsuite
