+++ setup.cfg
@@ -39,7 +39,7 @@ install_requires =
     gwpy >= 2.1.1
     healpy
     htcondor >= 9.0.6
-    lalsuite >= 7.5.1.dev20220614
+    lalsuite >= 7.6
     lintegrate >= 0.1.7
     matplotlib
     numba >= 0.5.0
