### 2022-08-08 2621546
Swap out lalapps for lalpulsar where required
