+++ docs/comparisons/scripts/comparitors.py
@@ -1,5 +1,5 @@
 """
-Output various comparitors between lalapps_pulsar_parameter_estimation_nested
+Output various comparitors between lalpulsar_parameter_estimation_nested
 runs and cwinpy runs
 """
 
@@ -21,7 +21,7 @@ FILETEXT = """\
    :widths: auto
    :header: "Method", ":math:`\\\\ln{{(Z)}}`", ":math:`\\\\ln{{(Z)}}` noise", ":math:`\\\\ln{{}}` Odds"
 
-   "``lalapps_pulsar_parameter_estimation_nested``", "{0:.3f}", "{1:.3f}", "{2:.3f}±{3:.3f}"
+   "``lalpulsar_parameter_estimation_nested``", "{0:.3f}", "{1:.3f}", "{2:.3f}±{3:.3f}"
    "``cwinpy_pe``", "{4:.3f}", "{5:.3f}", "{6:.3f}±{7:.3f}"
    "``cwinpy_pe`` (grid)", "{8}", "", "{9}"
 
@@ -29,7 +29,7 @@ FILETEXT = """\
    :widths: auto
    :header: "Method", ":math:`h_0`", ":math:`\\\\phi_0` (rad)", ":math:`\\\\psi` (rad)", ":math:`\\\\cos{{\\\\iota}}`"
 
-   "``lalapps_pulsar_parameter_estimation_nested``", "{10:.2f}±{11:.2f}×10\ :sup:`{12:d}`", "{13:.2f}±{14:.2f}", "{15:.2f}±{16:.2f}", "{17:.2f}±{18:.2f}"
+   "``lalpulsar_parameter_estimation_nested``", "{10:.2f}±{11:.2f}×10\ :sup:`{12:d}`", "{13:.2f}±{14:.2f}", "{15:.2f}±{16:.2f}", "{17:.2f}±{18:.2f}"
    "{19:d}% credible intervals", "[{20:.2f}, {21:.2f}]×10\ :sup:`{22:d}`", "[{23:.2f}, {24:.2f}]", "[{25:.2f}, {26:.2f}]", "[{27:.2f}, {28:.2f}]"
    "``cwinpy_pe``", "{29:.2f}±{30:.2f}×10\ :sup:`{31:d}`", "{32:.2f}±{33:.2f}", "{34:.2f}±{35:.2f}", "{36:.2f}±{37:.2f}"
    "{38:d}% credible intervals", "[{39:.2f}, {40:.2f}]×10\ :sup:`{41:d}`", "[{42:.2f}, {43:.2f}]", "[{44:.2f}, {45:.2f}]", "[{46:.2f}, {47:.2f}]"
@@ -38,7 +38,7 @@ FILETEXT = """\
    :widths: auto
    :header: "Method", ":math:`h_0`", ":math:`\\\\phi_0` (rad)", ":math:`\\\\psi` (rad)", ":math:`\\\\cos{{\\\\iota}}`", ":math:`\\\\ln{{(L)}}` max"
 
-   "``lalapps_pulsar_parameter_estimation_nested``", "{48:.2f}×10\ :sup:`{49:d}`", "{50:.2f}", "{51:.2f}", "{52:.2f}", "{53:.2f}"
+   "``lalpulsar_parameter_estimation_nested``", "{48:.2f}×10\ :sup:`{49:d}`", "{50:.2f}", "{51:.2f}", "{52:.2f}", "{53:.2f}"
    "``cwinpy_pe``", "{54:.2f}×10\ :sup:`{55:d}`", "{56:.2f}", "{57:.2f}", "{58:.2f}", "{59:.2f}"
 
 | Combined K-S test p-value: {60:.4f}
[…]
