+++ docs/comparisons/scripts/multi_detector_O1_data.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 """
-Compare cwinpy with lalapps_pulsar_parameter_estimation_nested for O1 data
+Compare cwinpy with lalpulsar_parameter_estimation_nested for O1 data
 from two detectors (H1 and L1). This uses data for the Crab pulsar.
 """
 
@@ -49,7 +49,7 @@ psirange = [0.0, np.pi / 2.0]
 cosiotarange = [-1.0, 1.0]
 h0range = [0.0, 1e-23]
 
-# set prior for lalapps_pulsar_parameter_estimation_nested
+# set prior for lalpulsar_parameter_estimation_nested
 priorfile = os.path.join(outdir, "{}_prior.txt".format(label))
 priorcontent = """H0 uniform {} {}
 PHI0 uniform {} {}
@@ -72,7 +72,7 @@ priors["cosiota"] = Uniform(
     cosiotarange[0], cosiotarange[1], "cosiota", latex_label=r"$\cos{\iota}$"
 )
 
-# run lalapps_pulsar_parameter_estimation_nested
+# run lalpulsar_parameter_estimation_nested
 try:
     execpath = os.environ["CONDA_PREFIX"]
 except KeyError:
@@ -82,7 +82,7 @@ except KeyError:
 
 execpath = os.path.join(execpath, "bin")
 
-lppen = os.path.join(execpath, "lalapps_pulsar_parameter_estimation_nested")
+lppen = os.path.join(execpath, "lalpulsar_parameter_estimation_nested")
 n2p = os.path.join(execpath, "lalinference_nest2pos")
 
 Nlive = 1000  # number of nested sampling live points
@@ -201,7 +201,7 @@ comparisons(label, outdir, grid, priors, cred=0.9)
 
 # create results plot
 allresults = {
-    "lalapps_pulsar_parameter_estimation_nested": outpost,
+    "lalpulsar_parameter_estimation_nested": outpost,
     "cwinpy_pe": result,
     "cwinpy_pe (grid)": grid,
 }
