+++ docs/comparisons/single-detector-software-injection-two-harmonics.txt
@@ -1,7 +1,7 @@
 Single detector, software injection, two harmonics
 ==================================================
 
-Here we compare ``lalapps_pulsar_parameter_estimation_nested`` with ``cwinpy`` in the case of
+Here we compare ``lalpulsar_parameter_estimation_nested`` with ``cwinpy`` in the case of
 simulated Gaussian noise from a single detector (H1 in this case) containing a software injected
 signal with two harmonics: one at the source rotation frequency and one at twice the rotation
 frequency. The parameters being estimated are :math:`C_{21}`, :math:`C_{22}`, :math:`\Phi_{21}`,
