### 2022-08-08 f57dca5
skyshifting.rst: minor docs updates
