### 2022-08-17 ea5871c
Update required bilby version
