+++ setup.cfg
@@ -32,7 +32,7 @@ setup_requires =
 install_requires =
     appdirs
     astropy
-    bilby >= 1.1.4
+    bilby >= 1.2.0
     configargparse
     corner >= 2.2.1
     gwosc >= 0.5.4
