### 2022-08-18 14b29df
likelihood.py: fix sign on Gaussian likelihood
