### 2022-08-18 24151b6
roq.py: minor fixes
