### 2022-08-25 #4568 @GregoryAshton
@mattpitkin I've now been able to install and run the software without a hitch.

My only remaining comment is about the tests. When I run `pytest` on the current master, I get

```python
============================================================================================== test session starts ==============================================================================================
rootdir: /home/greg/cwinpy
plugins: ligo.skymap-1.0.2
collected 0 items / 1 error

==================================================================================================== ERRORS =====================================================================================================
_________________________________________________________________________________________ ERROR collecting test session _________________________________________________________________________________________
../anaconda3/envs/cw/lib/python3.9/importlib/__init__.py:127: in import_module
<frozen importlib._bootstrap>:1030: in _gcd_import
    ???
<frozen importlib._bootstrap>:1007: in _find_and_load
    ???
<frozen importlib._bootstrap>:972: in _find_and_load_unlocked
    ???
<frozen importlib._bootstrap>:228: in _call_with_frames_removed
    ???
<frozen importlib._bootstrap>:1030: in _gcd_import
    ???
<frozen importlib._bootstrap>:1007: in _find_and_load
    ???
<frozen importlib._bootstrap>:972: in _find_and_load_unlocked
    ???
<frozen importlib._bootstrap>:228: in _call_with_frames_removed
    ???
<frozen importlib._bootstrap>:1030: in _gcd_import
    ???
<frozen importlib._bootstrap>:1007: in _find_and_load
    ???
<frozen importlib._bootstrap>:986: in _find_and_load_unlocked
    ???
<frozen importlib._bootstrap>:680: in _load_unlocked
    ???
<frozen importlib._bootstrap_external>:850: in exec_module
    ???
<frozen importlib._bootstrap>:228: in _call_with_frames_removed
    ???
cwinpy/__init__.py:8: in <module>
    from .knope import skyshift
cwinpy/knope/__init__.py:1: in <module>
    from .knope import knope, knope_pipeline
cwinpy/knope/knope.py:7: in <module>
    from ..heterodyne.heterodyne import (
cwinpy/heterodyne/__init__.py:1: in <module>
    from .base import *
cwinpy/heterodyne/base.py:37: in <module>
    from .fastheterodyne import fast_heterodyne
E   ModuleNotFoundError: No module named 'cwinpy.heterodyne.fastheterodyne'
=============================================================================================== warnings summary ================================================================================================
cwinpy/utils.py:238
  /home/greg/cwinpy/cwinpy/utils.py:238: DeprecationWarning: invalid escape sequence \

cwinpy/utils.py:277
  /home/greg/cwinpy/cwinpy/utils.py:277: DeprecationWarning: invalid escape sequence \
    """

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
============================================================================================ short test summary info ============================================================================================
ERROR  - ModuleNotFoundError: No module named 'cwinpy.heterodyne.fastheterodyne'
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 1 error during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!

```

When I look to see if the github actions are passing they are green, but looking in detail I see "Error" for all lines after "Checkout" https://github.com/cwinpy/cwinpy/runs/7875416995?check_suite_focus=true. I'm not sure if this is an access issue or not though.
