### 2022-09-07 4326d1f
setup.cfg: bump the required lalsuite version to 7.8
