### 2022-09-08 #4568 @mattpitkin
@GregoryAshton regarding running the tests, you'll need to pip install in editable mode to get these to work. I've got this noted in the documentation here https://cwinpy.readthedocs.io/en/latest/installation.html#testing.
