### 2022-09-09 e928be9
setup.cfg: update required lalsuite version
