### 2022-09-09 903f92a
data.py: add function to generate ROQ for a HeterodynedData object
