### 2022-09-09 bae9b80
likelihood.py: add in ability to use ROQ for generating the likelihood - currently this will not use a numba-ified likelihood
