+++ cwinpy/pe/likelihood.py
@@ -4,6 +4,7 @@ Classes providing likelihood functions.
 
 import re
 from copy import deepcopy
+from warnings import warn
 
 import bilby
 import lal
@@ -11,6 +12,7 @@ import numpy as np
 from astropy.time import Time
 from numba import jit, types
 from numba.typed import Dict as numbadict
+from scipy.linalg import lu_solve
 
 from ..data import HeterodynedData, MultiHeterodynedData
 from ..parfile import EPOCHPARS, PPUNITS, TEMPOUNITS
@@ -46,8 +48,10 @@ class TargetedPulsarLikelihood(bilby.core.likelihood.Likelihood):
         initialised.
     likelihood: str
         A string setting which likelihood function to use. This can either be
-        'studentst' (the default), or 'gaussian' ('roq' should be added in the
-        future).
+        `"studentst"` (the default), or `"gaussian"`.
+    roq: bool
+        Set to True if wanting to use reduced order quadrature for generating
+        the likelihood. Defaults to False.
     numba: bool
         Boolean to set whether to use the `numba` JIT compiled version of the
         likelihood function.
@@ -161,7 +165,9 @@ class TargetedPulsarLikelihood(bilby.core.likelihood.Likelihood):
     # the parameters that are held as vectors
     VECTOR_PARAMS = ["F", "GLEP", "GLPH", "GLF0", "GLF1", "GLF2", "GLF0D", "GLTD", "FB"]
 
-    def __init__(self, data, priors, likelihood="studentst", numba=False):
+    def __init__(
+        self, data, priors, likelihood="studentst", roq=False, numba=False, **kwargs
+    ):
 
         super().__init__(dict())  # initialise likelihood class
 
@@ -193,8 +199,11 @@ class TargetedPulsarLikelihood(bilby.core.likelihood.Likelihood):
         else:
             self.priors = priors
 
+        self.kwargs = kwargs.copy()
+
         # set the likelihood function
         self.likelihood = likelihood
+        self.roq = roq
         self._noise_log_likelihood = -np.inf  # initialise noise log likelihood
         self.numba = numba
 
@@ -255,14 +264,18 @@ class TargetedPulsarLikelihood(bilby.core.likelihood.Likelihood):
         # set up signal model classes
         self.models = []
         self.basepars = []
-        for het in self.data:
+        for i, het in enumerate(self.data):
+            times = het.times if not self.roq else self.__roq_all_nodes[i]
+            dt = 60 if len(times) == 1 else None
             self.models.append(
                 HeterodynedCWSimulator(
                     het.par,
                     het.detector,
-                    times=het.times,
+                    times=times,
+                    dt=dt,
                     earth_ephem=het.ephemearth,
                     sun_ephem=het.ephemsun,
+                    usetempo2=self.kwargs.get("usetempo2", False),
                 )
             )
             # copy of heterodyned parameters
@@ -297,6 +310,81 @@ class TargetedPulsarLikelihood(bilby.core.likelihood.Likelihood):
             self.__likelihood = "studentst"
         else:
             self.__likelihood = "gaussian"
+
+    @property
+    def roq(self):
+        """
+        A boolean defining whether to use ROQ for likelihood calculations.
+        """
+
+        return self.__roq
+
+    @roq.setter
+    def roq(self, roq):
+        if not isinstance(roq, bool):
+            raise TypeError("roq must be a boolean")
+
+        self.__roq = roq
+
+        self.__roq_all_nodes = []
+        self.__roq_all_real_node_indices = []
+        self.__roq_all_Bmat_lu_real = []
+        self.__roq_all_Dvec_real = []
+        self.__roq_all_imag_node_indices = []
[…]
