### 2022-09-09 6a74196
setup.cfg: move arby package from being a requirement to optional
