+++ setup.cfg
@@ -31,7 +31,6 @@ setup_requires =
     wheel
 install_requires =
     appdirs
-    arby
     astropy
     bilby >= 1.1.4
     configargparse
@@ -62,6 +61,7 @@ cwinpy =
 
 [options.extras_require]
 test =
+    arby
     codecov
     coverage
     coverage-badge
