### 2022-09-09 1efe22b
Add ability to use ROQ from the PE and knope pipelines
