+++ docs/knope/cwinpy_knope_pipeline.ini
@@ -510,3 +510,10 @@ priors = /home/username/analysis/priors
 # A flag to set whether to use the numba-enable likelihood function (defaults
 # to True)
 ;numba =
+
+# A flag to set whether to generate and use a reduced order quadrature for the
+# likelihood calculation (defaults to False)
+;roq =
+
+# A dictionary of any keyword arguments required by the ROQ generation
+;roq_kwargs =
