+++ docs/pe/cwinpy_pe_pipeline.ini
@@ -237,3 +237,10 @@ priors = /root/priors
 # A flag to set whether to use the numba-enable likelihood function (defaults
 # to True)
 ;numba =
+
+# A flag to set whether to generate and use a reduced order quadrature for the
+# likelihood calculation (defaults to False)
+;roq =
+
+# A dictionary of any keyword arguments required by the ROQ generation
+;roq_kwargs =
