### 2022-09-09 1528400
likelihood.py: fixed clashes when numba and roq flags are defined
