### 2022-09-09 9b3738c
make prior and likelihood not be included in plots by default
