### 2022-09-12 c5f3484
Allow using ROQ generation in verbose mode showing progress bar
