### 2022-09-15 df0870d
likelihood.py: fix noise log likelihood calculation for gaussian likelihood
