### 2022-09-15 12eecef
Fix Gaussian likelihood when using ROQ
