### 2022-09-15 063aefd
Start adding test suite for ROQ
