+++ cwinpy/test/test_roq.py
@@ -1,0 +1,326 @@
+"""
+Test script for ROQ usage.
+"""
+
+import copy
+
+import numpy as np
+import pytest
+from astropy.coordinates import SkyCoord
+from bilby.core.prior import PriorDict, Sine, Uniform
+from cwinpy.data import HeterodynedData
+from cwinpy.parfile import PulsarParameters
+from cwinpy.pe.roq import GenerateROQ
+from cwinpy.utils import logfactorial
+
+
+def full_log_likelihood(model, data, like="studentst", sigma=1.0):
+    """
+    Calculate the likelihood for the full data set.
+    """
+
+    dd = np.vdot(data, data).real
+    mm = np.vdot(model, model).real
+    dm = np.vdot(data, model).real
+
+    chisq = (dd - 2.0 * dm + mm) / sigma**2
+
+    if like == "studentst":
+        return (
+            logfactorial(len(data) - 1)
+            - np.log(2.0)
+            - len(data) * (np.log(np.pi * chisq))
+        )
+    else:
+        N = len(data) if data.dtype == complex else len(data) / 2
+        return -0.5 * chisq - N * np.log(2.0 * np.pi * sigma**2)
+
+
+class TestGenericModelROQ:
+    """
+    Test the reduced order quadrature for a generic model function.
+    """
+
+    @staticmethod
+    def generic_complex_model(t, A=0, phi0=0, m=0, c=0):
+        """
+        A (complex) sinusoid and a straight line model.
+        """
+
+        return A * np.exp(2.0 * np.pi * 1.4 * t * 1j + phi0) + m * t + c
+
+    @staticmethod
+    def generic_real_model(t, A=0, phi0=0, m=0, c=0):
+        """
+        A sinusoid and a straight line model.
+        """
+
+        return A * np.sin(2.0 * np.pi * 1.4 * t + phi0) + m * t + c
+
+    @classmethod
+    def setup_class(cls):
+        # create some complex data using generic_complex_model
+        cls.A_true = 1.2
+        cls.phi0_true = 2.3
+        cls.m_true = 0.7
+        cls.c_true = -1.2
+
+        cls.N = 100
+        cls.times = np.linspace(0, 5, cls.N)
+
+        cls.comp_model = cls.generic_complex_model(
+            cls.times,
+            A=cls.A_true,
+            phi0=cls.phi0_true,
+            m=cls.m_true,
+            c=cls.c_true,
+        )
+
+        cls.sigma = 1.5
+
+        cls.comp_data = cls.comp_model + (
+            np.random.normal(loc=0.0, scale=cls.sigma, size=cls.N)
+            + 1j * np.random.normal(loc=0.0, scale=cls.sigma, size=cls.N)
+        )
+
+        cls.real_model = cls.generic_real_model(
+            cls.times,
+            A=cls.A_true,
+            phi0=cls.phi0_true,
+            m=cls.m_true,
+            c=cls.c_true,
+        )
+
+        cls.real_data = cls.real_model + np.random.normal(
+            loc=0.0, scale=cls.sigma, size=cls.N
+        )
+
+        cls.priors = PriorDict()
+        cls.priors["A"] = Uniform(0, 10, name="A")
+        cls.priors["phi0"] = Uniform(0, 2.0 * np.pi, name="phi0")
+        cls.priors["m"] = Uniform(-5, 5, name="m")
+        cls.priors["c"] = Uniform(-5, 5, name="c")
+
+    def test_exceptions(self):
+        """
+        Test exceptions for GenerateROQ class.
+        """
+
+        # x-value exceptions
+        with pytest.raises(TypeError):
+            GenerateROQ(self.real_data, "sdjf", self.priors)
+
[…]
