### 2022-09-22 19805b5
test_roq.py: add a further test for the ROQ
