### 2022-09-25 0cab57d
test_roq.py: add more tests of ROQ code
