+++ cwinpy/pe/likelihood.py
@@ -265,7 +265,7 @@ class TargetedPulsarLikelihood(bilby.core.likelihood.Likelihood):
         self.models = []
         self.basepars = []
         for i, het in enumerate(self.data):
-            times = het.times if not self.roq else self.__roq_all_nodes[i]
+            times = het.times if not self.roq else self._roq_all_nodes[i]
             dt = het.times.value[1] - het.times.value[0]
             self.models.append(
                 HeterodynedCWSimulator(
@@ -328,16 +328,16 @@ class TargetedPulsarLikelihood(bilby.core.likelihood.Likelihood):
         if not self.__roq:
             return
 
-        self.__roq_all_nodes = []
-        self.__roq_all_real_node_indices = []
-        self.__roq_all_Bmat_lu_real = []
-        self.__roq_all_Dvec_real = []
-        self.__roq_all_imag_node_indices = []
-        self.__roq_all_Bmat_lu_imag = []
-        self.__roq_all_Dvec_imag = []
-        self.__roq_all_model2_node_indices = []
-        self.__roq_all_B2mat_lu = []
-        self.__roq_all_Bvec = []
+        self._roq_all_nodes = []
+        self._roq_all_real_node_indices = []
+        self._roq_all_Bmat_lu_real = []
+        self._roq_all_Dvec_real = []
+        self._roq_all_imag_node_indices = []
+        self._roq_all_Bmat_lu_imag = []
+        self._roq_all_Dvec_imag = []
+        self._roq_all_model2_node_indices = []
+        self._roq_all_B2mat_lu = []
+        self._roq_all_Bvec = []
 
         roqkwargs = self.kwargs.copy()
         roqkwargs["likelihood"] = self.likelihood
@@ -356,7 +356,7 @@ class TargetedPulsarLikelihood(bilby.core.likelihood.Likelihood):
                 self.__roq = False
                 return
 
-            self.__roq_all_nodes.append(np.concatenate([r._x_nodes for r in roqchunks]))
+            self._roq_all_nodes.append(np.concatenate([r._x_nodes for r in roqchunks]))
 
             # set node indices
             real_node_indices = []
@@ -382,15 +382,15 @@ class TargetedPulsarLikelihood(bilby.core.likelihood.Likelihood):
                 Dvec_imag.append(r._Dvec_imag)
                 Bvec.append(r._Bvec)
 
-            self.__roq_all_real_node_indices.append(real_node_indices)
-            self.__roq_all_imag_node_indices.append(imag_node_indices)
-            self.__roq_all_model2_node_indices.append(model2_node_indices)
-            self.__roq_all_Bmat_lu_real.append(Bmat_lu_real)
-            self.__roq_all_Bmat_lu_imag.append(Bmat_lu_imag)
-            self.__roq_all_Dvec_real.append(Dvec_real)
-            self.__roq_all_Dvec_imag.append(Dvec_imag)
-            self.__roq_all_B2mat_lu.append(B2mat_lu)
-            self.__roq_all_Bvec.append(Bvec)
+            self._roq_all_real_node_indices.append(real_node_indices)
+            self._roq_all_imag_node_indices.append(imag_node_indices)
+            self._roq_all_model2_node_indices.append(model2_node_indices)
+            self._roq_all_Bmat_lu_real.append(Bmat_lu_real)
+            self._roq_all_Bmat_lu_imag.append(Bmat_lu_imag)
+            self._roq_all_Dvec_real.append(Dvec_real)
+            self._roq_all_Dvec_imag.append(Dvec_imag)
+            self._roq_all_B2mat_lu.append(B2mat_lu)
+            self._roq_all_Bvec.append(Bvec)
 
     def dot_products(self):
         """
[…]
