### 2022-09-28 ed04947
Fixing nested references
