+++ paper/paper.bib
@@ -131,7 +131,7 @@
 }
 
 @misc{tempo2manual,
-  author       = {{Hobbs}, G. and {Edwards}, R.},
+  author       = {{Hobbs}, G.~B. and {Edwards}, R.},
   title        = {{TEMPO2} user manual},
   url          = {https://www.jb.man.ac.uk/~pulsar/Resources/tempo2_manual.pdf},
   year         = {2006},
