### 2022-09-28 #4568 @dfm
@GregoryAshton, @ColmTalbot — Thank you for your volunteered time and constructive reviews!! We couldn't do this without you and I really appreciate the time you gave us! 🚀

@mattpitkin — I've opened a PR with some minor edits to the paper. After merging or responding to that, here are the final steps that I'll need from you:

1. Take one last read through the manuscript to make sure that you're happy with it (it's harder to make changes later!), especially your and affiliation. I've taken a pass and it looks good to me!
