+++ cwinpy/test/test_roq.py
@@ -3,16 +3,21 @@ Test script for ROQ usage.
 """
 
 import copy
+import os
+import shutil
+import subprocess as sp
 
 import numpy as np
 import pytest
 from astropy.coordinates import SkyCoord
 from bilby.core.prior import PriorDict, Sine, Uniform
 from cwinpy.data import HeterodynedData, MultiHeterodynedData
+from cwinpy.heterodyne import heterodyne
 from cwinpy.parfile import PulsarParameters
+from cwinpy.pe import pe
 from cwinpy.pe.likelihood import TargetedPulsarLikelihood
 from cwinpy.pe.roq import GenerateROQ
-from cwinpy.utils import logfactorial
+from cwinpy.utils import LAL_EPHEMERIS_URL, download_ephemeris_file, logfactorial
 
 
 def full_log_likelihood(model, data, like="studentst", sigma=1.0):
@@ -436,3 +441,202 @@ class TestHeterodynedCWModelROQ:
             llr[i] = like_roq.log_likelihood()
 
         assert np.all(np.abs(np.exp(llo - llo.max()) - np.exp(llr - llr.max())) < 1e-3)
+
+
+class TestROQFrequency(object):
+    """
+    Test the generation of an ROQ over frequency parameters.
+    """
+
+    @classmethod
+    def setup_class(cls):
+        # create some fake data frames using lalpulsar_Makefakedata_v5
+        mfd = shutil.which("lalpulsar_Makefakedata_v5")
+
+        cls.fakedatadir = "testing_fake_frame_cache"
+        cls.fakedatadetector = "H1"
+        cls.fakedatachannel = f"{cls.fakedatadetector}:FAKE_DATA"
+
+        cls.fakedatastart = 1000000000
+        cls.fakedataduration = 86400
+
+        os.makedirs(cls.fakedatadir, exist_ok=True)
+
+        cls.fakedatabandwidth = 8  # Hz
+        sqrtSn = 1e-24  # noise amplitude spectral density
+        cls.fakedataname = "FAKEDATA"
+
+        # Create two pulsars to inject: one isolated and one binary
+        cls.fakepulsarpar = []
+
+        # requirements for Makefakedata pulsar input files
+        pulsarstr = """\
+Alpha = {alpha}
+Delta = {delta}
+Freq = {f0}
+f1dot = {f1}
+f2dot = {f2}
+refTime = {pepoch}
+h0 = {h0}
+cosi = {cosi}
+psi = {psi}
+phi0 = {phi0}
+"""
+
+        # pulsar parameters
+        f0 = 6.9456 / 2.0  # source rotation frequency (Hz)
+        f1 = -9.87654e-11 / 2.0  # source rotational frequency derivative (Hz/s)
+        f2 = 2.34134e-18 / 2.0  # second frequency derivative (Hz/s^2)
+        alpha = 0.0  # source right ascension (rads)
+        delta = 0.5  # source declination (rads)
+        pepoch = 1000000000  # frequency epoch (GPS)
+
+        # GW parameters
+        h0 = 3.0e-24  # GW amplitude
+        phi0 = 1.0  # GW initial phase (rads)
+        cosiota = 0.1  # cosine of inclination angle
+        psi = 0.5  # GW polarisation angle (rads)
+
+        mfddic = {
+            "alpha": alpha,
+            "delta": delta,
+            "f0": 2 * f0,
+            "f1": 2 * f1,
+            "f2": 2 * f2,
+            "pepoch": pepoch,
+            "h0": h0,
+            "cosi": cosiota,
+            "psi": psi,
+            "phi0": phi0,
+        }
+
+        cls.fakepulsarpar = PulsarParameters()
+        cls.fakepulsarpar["PSRJ"] = "J0000+0000"
[…]
