### 2022-09-28 #4568 @mattpitkin
@dfm - I've made some additional minor changes to the paper, but it looks good now. One minor thing is that Greg and Colm appear listed twice (once with fully lowercase Github usernames and again with camel case usernames) in the reviewers section of the paper.

I get a new release of the software drafted now.
