### 2022-09-28 #4568 @mattpitkin
> Create an archived release of that version of the software (using Zenodo or something similar). Please make sure that the metadata (title and author list) exactly match the paper. Then report the DOI of the release back to this thread.

The Zenodo DOI is [10.5281/zenodo.7121400](https://doi.org/10.5281/zenodo.7121400)
