### 2022-09-29 #4568 @dfm
@openjournals/dev — There's an issue with this review where the reviewers are listed twice. I think this happened because their usernames were initially saved as all lowercase and I had to re-add them to get the usernames saved properly. I updated the first comment in this thread to remove the duplicate usernames, but the generated PDF still has the lowercase usernames listed. Should we just accept and fix this post? Or is there something else we can do to fix this in advance? Thanks!
